Revert "Filter available distributions using hash declarations from constraints files (#10962)"
This reverts commit 0c284520c6d068cb25ac89d9dbee0456c2eba23a.
